Initial Commit
The first commit of this extension.  Again this is very specific to a instance of Service Now.  I'm not sure if the same issues plague other instances.  

This aims to reduce the size of the comments that are in the activity stream.  In that instance, the comment max length defaults to 4000 characters.  I've been told it was a bug, so decided to address it using a chrome extension.  The chrome extension simply finds the comments by CSS class name, then resizes all those with a client height greater than the value entered in the chrome extension. 

Also there is too much white space on the bottom of every ticket, so it resizes that element as well. 

The chrome extension popup is not very visually appealing.  Probably should address that next.